+++ safire/include/safire/ecs.h
@@ -12,16 +12,18 @@ typedef struct SFRentity                SFRentity_t;
 typedef struct SFRcomponent             SFRcomponent_t;
 typedef struct SFRscene                 SFRscene_t;
 
-
-
+typedef uint64_t                        SFRuuid_t;
 
 SAFIRE_API void             sfr_ecs_init(SFRscene_t** scenes, uint32_t scenes_count);
 SAFIRE_API SFRecs_t*        sfr_ecs_instance();
 SAFIRE_API void             sfr_ecs_free();
-SAFIRE_API void             sfr_ecs_clear_stack();
+SAFIRE_API void             sfr_ecs_clear_stack();                                        // frees all entities & components from the stack
 
-SAFIRE_API void             sfr_ecs_update(float delta_time);
-SAFIRE_API void             sfr_ecs_late_update(float late_delta_time);
+SAFIRE_API void             sfr_ecs_update(float delta_time);                             // calls all the components update loops that have it
+SAFIRE_API void             sfr_ecs_late_update(float late_delta_time);                   // calls al the components late update loops that have it
+
+SAFIRE_API void             sfr_ecs_erase_entity(uint32_t index);                         // removes the target entity from the stack
+SAFIRE_API void             sfr_ecs_erase_component(uint32_t index);                      // removes the target component from the stack
 
 SAFIRE_API void             sfr_ecs_load_scene(uint32_t id);
 
@@ -44,15 +46,19 @@ struct SFRscene {
   sfr_scene_free_data       free;
 };
 
-SAFIRE_API SFRscene_t*      sfr_scene(const char* name, sfr_scene_init start);
-SAFIRE_API uint32_t         sfr_find_scene(const char* name);
+SAFIRE_API SFRscene_t*      sfr_scene(const char* name, sfr_scene_init start);            // creates a scene instance
+SAFIRE_API uint32_t         sfr_find_scene(const char* name);                             // finds a scene instance so you can change the current scene
 
 
 
 
 struct SFRentity {
-  uint64_t                  uuid;
+  SFRuuid_t                 uuid;
   char*                     name;
+  char*                     tag;
+
+  SFRuuid_t*                components;
+  uint32_t                  components_count;
 };
 
 typedef void (*component_attach)        (SFRcomponent_t*);
@@ -61,9 +67,10 @@ typedef void (*component_late_update)   (SFRcomponent_t*, float);
 typedef void (*component_free)          (SFRcomponent_t*);
 
 struct SFRcomponent {
-  uint64_t                  uuid;
+  SFRuuid_t                 uuid;
   char*                     name;
   SFRentity_t*              owner;
+  void*                     data;
 
   component_attach attach;
   component_update update;
@@ -77,12 +84,18 @@ SAFIRE_API SFRcomponent_t*  sfr_ecs_push_component(SFRentity_t* entity, SFRcompo
 SAFIRE_API void             sfr_ecs_entity_free(SFRentity_t* entity);
 SAFIRE_API void             sfr_ecs_component_free(SFRcomponent_t* component);
 
-SAFIRE_API void             sfr_ecs_entity_target_free();
-SAFIRE_API void             sfr_ecs_component_target_free();
+SAFIRE_API void             sfr_ecs_entity_target_free(uint32_t id);
+SAFIRE_API void             sfr_ecs_component_target_free(uint32_t id);
 
 SAFIRE_API SFRentity_t*     sfr_ecs_find_entity_name(const char* name);
-SAFIRE_API SFRentity_t*     sfr_ecs_find_entity_uuid(uint64_t uuid);
-SAFIRE_API SFRcomponent_t** sfr_ecs_find_list_components_name(const char* name);
+SAFIRE_API SFRentity_t**    sfr_ecs_find_list_entities(const char* tag, uint32_t* entity_count);
+SAFIRE_API SFRentity_t*     sfr_ecs_find_entity_uuid(SFRuuid_t uuid);
+SAFIRE_API SFRcomponent_t*  sfr_ecs_find_component_uuid(SFRuuid_t uuid);
+SAFIRE_API SFRcomponent_t** sfr_ecs_find_list_components(const char* name, uint32_t* component_count);
+
+SAFIRE_API uint32_t         sfr_ecs_entity_find_index_name(uint32_t offset, const char* name);
+SAFIRE_API uint32_t         sfr_ecs_entity_find_index_uuid(uint32_t offset, SFRuuid_t uuid);
+SAFIRE_API uint32_t         sfr_ecs_component_find_index_uuid(uint32_t offset, SFRuuid_t uuid);