+++ safire/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * made code more cleaner *(VERSION 0.0.6.1)*
   * spacing and tabbing for functions and type defines in the header files
+* ecs implementation
+  * finding entities and components *(VERSION 0.0.6.2)*
 
 ## VERSION 0.0.6 (25/4/2022)